Extract the visualization code into its own module.
